third_party/blink/web_tests/platform/generic/fast/js/array-sort-exception-expected.txt

Test of array sort with toString() override that throws exception.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS catchArg is exceptionString
PASS successfullyParsed is true

TEST COMPLETE
